MY projects/diabetic_retinopathy_app/diabetic_retinopathy_app/requirements.txt
flask==2.0.1
flask-login==0.5.0
mysql-connector-python==8.0.26
tensorflow==2.6.0
opencv-python==4.5.3.56
numpy==1.19.5
python-dotenv==0.19.0
werkzeug==2.0.1